Rename login() to getLoggedInUser()
