+++ app/src/main/java/de/hbch/traewelling/api/models/auth/LoginCredentials.kt
@@ -3,6 +3,6 @@ package de.hbch.traewelling.api.models.auth
 import com.google.gson.annotations.SerializedName
 
 class LoginCredentials(
-    @SerializedName("login") val login: String,
+    @SerializedName("getLoggedInUser") val login: String,
     @SerializedName("password") val password: String
     )
